# Replace `stopping_criteria` logic for `generate_reply_HF`