@Panchovix 2023-06-22
Just tested and I can confirm it works with exllama_hf, nice work!