@oobabooga 2023-06-22
I'm not sure if using the streaming hack even when streaming is disabled is a good idea. It spawns a new thread, which I think may lead to higher VRAM usage.