@official-elinas 2023-06-23
@oobabooga I'm not noticing any difference in VRAM usage, just that my tk/s dropped in half.