@Panchovix 2023-06-23
I see same VRAM usage as well, speed is the same also.