@Yiximail 2023-06-23
> 

dropped in half? 

Can you provide your computer configuration and the usage of CPU, memory, GPU and VRAM when running?
I'm not sure what caused it, but it's possible that one of them has reached a bottleneck.

T^T