@oobabooga 2023-06-24
If we are going to enable streaming for the sake of checking stopping strings, it's easier to do that directly in the `generate_reply` function. This way, it works for all models.

This is done in https://github.com/oobabooga/text-generation-webui/pull/2847 in a simple way, so I'm closing this PR in favor of that one.